Merge improvements from Develop

+++ lib/aes/aesblock.cpp
@@ -177,7 +177,7 @@ void aesBlock::mixColumns() {
     for (uint8_t column = 0; column < 4; column++) {
         for (uint8_t row = 0; row < 4; row++) {
             a[row] = tempState[row][column];
-            b[row] = (tempState[row][column] << 1U);
+            b[row] = static_cast<uint8_t>(tempState[row][column] << 1U);
 
             if ((tempState[row][column] & 0x80) == 0x80) {
                 b[row] ^= 0x1B;
@@ -195,13 +195,13 @@ void aesBlock::shiftLeft() {
     for (uint32_t byteIndex = 0; byteIndex < 16; byteIndex++) {
         if (byteIndex < 15) {
             if ((state.asByte[byteIndex + 1] & 0x80) == 0x80) {
-                state.asByte[byteIndex] = (state.asByte[byteIndex] << 1U) + 1U;
+                state.asByte[byteIndex] = static_cast<uint8_t>((state.asByte[byteIndex] << 1U) + 1U);
 
             } else {
-                state.asByte[byteIndex] = (state.asByte[byteIndex] << 1U);
+                state.asByte[byteIndex] = static_cast<uint8_t>(state.asByte[byteIndex] << 1U);
             }
         } else {
-            state.asByte[byteIndex] = (state.asByte[byteIndex] << 1U);
+            state.asByte[byteIndex] = static_cast<uint8_t>(state.asByte[byteIndex] << 1U);
         }
     }
 }
